requirements.txt
Django==4.2
django-environ==0.10.0
Pillow==9.5.0
social-auth-app-django==4.0.0
celery==5.2.7
flower==1.2.0
braintree==4.19.0
weasyprint==58.1
django-crispy-forms==2.0
psycopg==3.1.8
python-jose==3.3.0
gunicorn==20.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
crispy-bootstrap4==2022.1
djoser==2.1.0
black==23.3.0
flake8==6.0.0
pre-commit==3.2.2
isort==5.12.0
